requirements.txt
awsebcli==3.12.1
Babel==2.5.3
blessed==1.14.2
botocore==1.8.36
cement==2.8.2
colorama==0.3.7
Django==2.0.1
django-classy-tags==0.8.0
django-compat==1.0.15
django-countries==5.0
django-fontawesome==0.3.1
django-phonenumber-field==2.0.0
django-taggit==0.22.2
django-taggit-templatetags2==1.6.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
pathspec==0.5.0
phonenumberslite==8.8.10
Pillow==5.0.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.46.0
